+++ docs/conf.py
@@ -23,7 +23,7 @@ copyright = "2021, David Montero Loaiza"
 author = "David Montero Loaiza"
 
 # The full version, including alpha/beta/rc tags
-release = "0.2.6"
+release = "0.3.1"
 
 # -- General configuration ---------------------------------------------------
 
@@ -39,6 +39,8 @@ extensions = [
     "sphinx_rtd_theme",
     "sphinx.ext.autosummary",
     "sphinx.ext.autosectionlabel",
+    "sphinx_copybutton",
+    "nbsphinx",
 ]
 
 autodoc_default_flags = ["members"]