+++ eemont/__init__.py
@@ -1,17 +1,7 @@
 """The eemont python package."""
 
-__version__ = "0.2.6"
+__version__ = "0.3.1"
 
-from . import (
-    dataframe,
-    eeDictionary,
-    eeList,
-    extra,
-    feature,
-    featurecollection,
-    geometry,
-    image,
-    imagecollection,
-    number,
-)
+from . import (app, dataframe, eeDictionary, eeList, extra, feature,
+               featurecollection, geometry, image, imagecollection, number)
 from .common import *