+++ eemont/image.py
@@ -1,21 +1,11 @@
 import warnings
 
 import ee
+import ee_extra
+import ee_extra.Spectral.core
 import requests
 
-from .common import (
-    _get_offset_params,
-    _get_scale_params,
-    _getCitation,
-    _getDOI,
-    _getSTAC,
-    _index,
-    _maskClouds,
-    _matchHistogram,
-    _panSharpen,
-    _preprocess,
-    _scale_STAC,
-)
+from .common import _matchHistogram, _panSharpen
 from .extending import extend
 
 
@@ -665,18 +655,18 @@ def index(
     alpha=0.1,
     slope=1.0,
     intercept=0.0,
+    gamma=1.0,
     kernel="RBF",
     sigma="0.5 * (a + b)",
     p=2.0,
     c=1.0,
     online=False,
+    drop=False,
 ):
     """Computes one or more spectral indices (indices are added as bands) for an image.
 
-    Warning
-    -------------
-    **Pending Deprecation:** The :code:`index()` method will no longer be available for
-    future versions. Please use :code:`spectralIndices()` instead.
+    .. deprecated:: 0.3.0
+       Use :func:`spectralIndices()` instead.
 
     Tip
     ----------
@@ -714,11 +704,13 @@ def index(
     nexp : float, default = 2.0
         Exponent used for GDVI.
     alpha : float, default = 0.1
-        Weighting coefficient  used for WDRVI.
+        Weighting coefficient used for WDRVI.
     slope : float, default = 1.0
         Soil line slope.
     intercept : float, default = 0.0
         Soil line intercept.
+    gamma : float, default = 1.0
+        Weighting coefficient used for ARVI.
     kernel : str, default = 'RBF'
         Kernel used for kernel indices.\n
         Available options:
@@ -735,10 +727,10 @@ def index(
         terms in the polynomial kernel. Used for kernel = 'poly'. This must be greater
         than or equal to 0.
     online : boolean, default = False
-        Wheter to retrieve the most recent list of indices directly from the GitHub
+        Whether to retrieve the most recent list of indices directly from the GitHub
         repository and not from the local copy.
-
-        .. versionadded:: 0.2.0
+    drop : boolean, default = True
+        Whether to drop all bands except the new spectral indices.
 
     Returns
     -------
@@ -781,11 +773,11 @@ def index(
     >>> S2.index('all')
     """
     warnings.warn(
-        "index() will be deprecated in future versions, please use spectralIndices() instead",
-        PendingDeprecationWarning,
+        "index() is deprecated, please use spectralIndices() instead",
+        DeprecationWarning,
     )
 
-    return _index(
+    return ee_extra.Spectral.core.spectralIndices(
         self,
         index,
         G,
@@ -797,11 +789,13 @@ def index(
         alpha,
         slope,
         intercept,
+        gamma,
         kernel,
         sigma,
         p,
         c,
         online,
+        drop,
     )
 
 
@@ -818,11 +812,13 @@ def spectralIndices(
     alpha=0.1,
     slope=1.0,
     intercept=0.0,
+    gamma=1.0,
     kernel="RBF",
     sigma="0.5 * (a + b)",
     p=2.0,
     c=1.0,
     online=False,
+    drop=False,
 ):
     """Computes one or more spectral indices (indices are added as bands) for an image
     from the Awesome List of Spectral Indices.
@@ -863,11 +859,13 @@ def spectralIndices(
     nexp : float, default = 2.0
         Exponent used for GDVI.
     alpha : float, default = 0.1
-        Weighting coefficient  used for WDRVI.
+        Weighting coefficient used for WDRVI.
     slope : float, default = 1.0
         Soil line slope.
     intercept : float, default = 0.0
         Soil line intercept.
+    gamma : float, default = 1.0
+        Weighting coefficient used for ARVI.
     kernel : str, default = 'RBF'
         Kernel used for kernel indices.\n
         Available options:
@@ -884,8 +882,10 @@ def spectralIndices(
         terms in the polynomial kernel. Used for kernel = 'poly'. This must be greater
         than or equal to 0.
     online : boolean, default = False
-        Wheter to retrieve the most recent list of indices directly from the GitHub
+        Whether to retrieve the most recent list of indices directly from the GitHub
         repository and not from the local copy.
+    drop : boolean, default = True
+        Whether to drop all bands except the new spectral indices.
 
     Returns
     -------
@@ -927,7 +927,7 @@ def spectralIndices(
 
     >>> S2.spectralIndices('all')
     """
-    return _index(
+    return ee_extra.Spectral.core.spectralIndices(
         self,
         index,
         G,
@@ -939,11 +939,13 @@ def spectralIndices(
         alpha,
         slope,
         intercept,
+        gamma,
         kernel,
         sigma,
         p,
         c,
         online,
+        drop,
     )
 
 
@@ -1024,7 +1026,7 @@ def maskClouds(
     ...     .first()
     ...     .maskClouds(prob = 75,buffer = 300,cdi = -0.5))
     """
-    return _maskClouds(
+    return ee_extra.QA.clouds.maskClouds(
         self,
         method,
         prob,
@@ -1042,10 +1044,8 @@ def maskClouds(
 def scale(self):
     """Scales bands on an image.
 
-    Warning
-    -------------
-    **Pending Deprecation:** The :code:`scale()` method will no longer be available for
-    future versions. Please use :code:`scaleAndOffset()` instead.
+    .. deprecated:: 0.3.0
+       Use :func:`scaleAndOffset()` instead.
 
     Tip
     ----------
@@ -1070,11 +1070,11 @@ def scale(self):
     >>> S2 = ee.ImageCollection('COPERNICUS/S2_SR').first().scale()
     """
     warnings.warn(
-        "scale() will be deprecated in future versions, please use scaleAndOffset() instead",
-        PendingDeprecationWarning,
+        "scale() is deprecated, please use scaleAndOffset() instead",
+        DeprecationWarning,
     )
 
-    return _scale_STAC(self)
+    return ee_extra.STAC.core.scaleAndOffset(self)
 
 
 @extend(ee.image.Image)
@@ -1115,7 +1115,7 @@ def getScaleParams(self):
      'QC_Day': 1.0,
      'QC_Night': 1.0}
     """
-    return _get_scale_params(self)
+    return ee_extra.STAC.core.getScaleParams(self)
 
 
 @extend(ee.image.Image)
@@ -1156,7 +1156,7 @@ def getOffsetParams(self):
      'QC_Day': 0.0,
      'QC_Night': 0.0}
     """
-    return _get_offset_params(self)
+    return ee_extra.STAC.core.getOffsetParams(self)
 
 
 @extend(ee.image.Image)
@@ -1190,7 +1190,7 @@ def scaleAndOffset(self):
     >>> ee.Initialize()
     >>> S2 = ee.ImageCollection('COPERNICUS/S2_SR').first().scaleAndOffset()
     """
-    return _scale_STAC(self)
+    return ee_extra.STAC.core.scaleAndOffset(self)
 
 
 @extend(ee.image.Image)
@@ -1228,7 +1228,7 @@ def preprocess(self, **kwargs):
     >>> ee.Initialize()
     >>> S2 = ee.ImageCollection('COPERNICUS/S2_SR').first().preprocess()
     """
-    return _preprocess(self, **kwargs)
+    return ee_extra.QA.pipelines.preprocess(self, **kwargs)
 
 
 @extend(ee.image.Image)
@@ -1259,7 +1259,7 @@ def getSTAC(self):
      'gee:type': 'image_collection',
      ...}
     """
-    return _getSTAC(self)
+    return ee_extra.STAC.core.getSTAC(self)
 
 
 @extend(ee.image.Image)
@@ -1288,7 +1288,7 @@ def getDOI(self):
     >>> ee.ImageCollection('NASA/GPM_L3/IMERG_V06').first().getDOI()
     '10.5067/GPM/IMERG/3B-HH/06'
     """
-    return _getDOI(self)
+    return ee_extra.STAC.core.getDOI(self)
 
 
 @extend(ee.image.Image)
@@ -1321,7 +1321,7 @@ def getCitation(self):
     Accessed: [Data Access Date],
     [doi:10.5067/GPM/IMERG/3B-HH/06](https://doi.org/10.5067/GPM/IMERG/3B-HH/06)'
     """
-    return _getCitation(self)
+    return ee_extra.STAC.core.getCitation(self)
 
 
 @extend(ee.image.Image)
@@ -1405,3 +1405,65 @@ def matchHistogram(self, target, bands, geometry=None, maxBuckets=256):
     >>> matched = source.matchHistogram(target, bands)
     """
     return _matchHistogram(self, target, bands, geometry, maxBuckets)
+
+
+@extend(ee.image.Image)
+def tasseledCap(self):
+    """Calculates tasseled cap brightness, wetness, and greenness components.
+
+    Tasseled cap transformations are applied using coefficients published for these
+    supported platforms:
+
+    * Sentinel-2 MSI Level 1C [1]_
+    * Landsat 8 OLI TOA [2]_
+    * Landsat 7 ETM+ TOA [3]_
+    * Landsat 5 TM Raw DN [4]_
+    * Landsat 4 TM Raw DN [5]_
+    * Landsat 4 TM Surface Reflectance [6]_
+    * MODIS NBAR [7]_
+
+    Parameters
+    ----------
+    self : ee.Image
+        Image to calculate tasseled cap components for. Must belong to a supported
+        platform.
+
+    Returns
+    -------
+    ee.Image
+        Image with the tasseled cap components as new bands.
+
+    References
+    ----------
+    .. [1] Shi, T., & Xu, H. (2019). Derivation of Tasseled Cap Transformation
+        Coefficients for Sentinel-2 MSI At-Sensor Reflectance Data. IEEE Journal
+        of Selected Topics in Applied Earth Observations and Remote Sensing, 1–11.
+        doi:10.1109/jstars.2019.2938388
+    .. [2] Baig, M.H.A., Zhang, L., Shuai, T. and Tong, Q., 2014. Derivation of a
+        tasselled cap transformation based on Landsat 8 at-satellite reflectance.
+        Remote Sensing Letters, 5(5), pp.423-431.
+    .. [3] Huang, C., Wylie, B., Yang, L., Homer, C. and Zylstra, G., 2002.
+        Derivation of a tasselled cap transformation based on Landsat 7 at-satellite
+        reflectance. International journal of remote sensing, 23(8), pp.1741-1748.
+    .. [4] Crist, E.P., Laurin, R. and Cicone, R.C., 1986, September. Vegetation and
+        soils information contained in transformed Thematic Mapper data. In
+        Proceedings of IGARSS’86 symposium (pp. 1465-1470). Paris: European Space
+        Agency Publications Division.
+    .. [5] Crist, E.P. and Cicone, R.C., 1984. A physically-based transformation of
+        Thematic Mapper data---The TM Tasseled Cap. IEEE Transactions on Geoscience
+        and Remote sensing, (3), pp.256-263.
+    .. [6] Crist, E.P., 1985. A TM tasseled cap equivalent transformation for
+        reflectance factor data. Remote sensing of Environment, 17(3), pp.301-306.
+    .. [7] Lobser, S.E. and Cohen, W.B., 2007. MODIS tasselled cap: land cover
+        characteristics expressed through transformed MODIS data. International
+        Journal of Remote Sensing, 28(22), pp.5079-5101.
+
+    Examples
+    --------
+    >>> import ee, eemont
+    >>> ee.Authenticate()
+    >>> ee.Initialize()
+    >>> img = ee.Image("LANDSAT/LT05/C01/T1/LT05_044034_20081011")
+    >>> img = img.tasseledCap()
+    """
+    return ee_extra.Spectral.core.tasseledCap(self)